# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token pair issued. No role data is duplicated in the body — authorization claims live inside the JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResponseDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or deactivated account (uniform response)."
          }
        },
        "summary": "Exchange email + password for an access/refresh token pair.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fresh token pair; the presented refresh token is revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResponseDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unknown, expired, or already-rotated refresh token."
          }
        },
        "summary": "Rotate the refresh token and mint a new access token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session revoked; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current refresh token/session.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The authenticated identity attached to every request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/CurrentUserDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the verified session identity from the JWT claims.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/fleets": {
      "post": {
        "operationId": "FleetsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFleetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fleet created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "409": {
            "description": "Owner role slug unknown/inactive, or the owner user cannot be linked."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a fleet (platform path; optionally grants the owner an initial membership).",
        "tags": [
          "fleets"
        ]
      },
      "get": {
        "operationId": "FleetsController_findAll",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque nextCursor from the previous page (base64url of the last item id).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 20, max 100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of fleets (items + nextCursor).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FleetDto"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all fleets (platform path, cursor pagination).",
        "tags": [
          "fleets"
        ]
      }
    },
    "/fleets/mine": {
      "get": {
        "operationId": "FleetsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The caller's ACTIVE memberships — never another user's.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MyMembershipDto"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the caller's own ACTIVE fleet memberships (RLS tenant path).",
        "tags": [
          "fleets"
        ]
      }
    },
    "/fleets/{id}": {
      "get": {
        "operationId": "FleetsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The fleet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one fleet by id (platform path).",
        "tags": [
          "fleets"
        ]
      },
      "patch": {
        "operationId": "FleetsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFleetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated fleet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a fleet (platform path).",
        "tags": [
          "fleets"
        ]
      },
      "delete": {
        "operationId": "FleetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fleet deleted; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          },
          "409": {
            "description": "Fleet still referenced by buses/trips/bookings."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a fleet and its memberships (platform path, cascading).",
        "tags": [
          "fleets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Fleet API",
    "description": "Multi-tenant group bus transportation platform REST API.\n\n- **Authentication is owned by this API**: HS256 JWTs carry a custom `app_role` claim; refresh tokens rotate on every use and are stored hashed at rest.\n- **Every response is wrapped in a `{ statusCode, data }` envelope**; the `data` schema is shown inside each documented response.\n- **Fleet isolation is enforced by PostgreSQL RLS** as the final security boundary; cross-tenant lookups are indistinguishable from missing rows and return 404.\n- `fleetId` path parameters are **selectors, never proof of authorization** — access is verified against the caller's ACTIVE `FleetMember` row on every request.\n- Roles and permissions are **database-driven** (no hardcoded enums); only `super_admin` is a predefined system role. Platform routes (roles, permissions, users, fleets admin, audit-logs) are privileged: super admin only and audited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "HS256 access JWT carrying {sub, email, app_role, authVersion, sessionId}.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque rotating refresh token (7 days). Present it to /auth/refresh; never store it client-side in localStorage.",
            "example": "b3BhcXVlLXJlZnJlc2gtdG9rZW4"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@bus.local",
            "format": "email",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!123",
            "format": "password",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token obtained from login/refresh. Single use — it rotates.",
            "example": "b3BhcXVlLXJlZnJlc2gtdG9rZW4",
            "maxLength": 512
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "admin@bus.local",
            "format": "email"
          },
          "appRole": {
            "type": "string",
            "description": "Global role slug resolved from the database (super_admin priority, \"user\" fallback).",
            "example": "super_admin"
          },
          "authVersion": {
            "type": "number",
            "description": "Bumped on security-sensitive changes; stale tokens are rejected."
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "description": "Session id — checked against live sessions on every request."
          }
        },
        "required": [
          "id",
          "email",
          "appRole",
          "authVersion",
          "sessionId"
        ]
      },
...
      "FleetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Fleet A"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "ownerId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateFleetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fleet A",
            "minLength": 1,
            "maxLength": 255
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "description": "User id of the fleet owner (must exist and be active)."
          },
          "ownerRoleSlug": {
            "type": "string",
            "example": "fleet-owner",
            "description": "Optional initial ACTIVE membership for the owner."
          }
        },
        "required": [
          "name",
          "ownerId"
        ]
      },
      "MyMembershipDto": {
        "type": "object",
        "properties": {
          "fleetId": {
            "type": "string",
            "format": "uuid"
          },
          "fleetName": {
            "type": "string",
            "example": "Fleet A"
          },
          "roleSlug": {
            "type": "string",
            "example": "fleet-operator",
            "description": "Slug of the role held in this fleet."
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          }
        },
        "required": [
          "fleetId",
          "fleetName",
          "roleSlug",
          "status"
        ]
      },
      "UpdateFleetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fleet A",
            "minLength": 1,
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}